requirements.txt
langchain
langchain-classic
langchain-openai
python-dotenv
